Higgs Module/navigation.py
import numpy as np

try:
    import cupy as cp
    CUPY_AVAILABLE = True
except ImportError:
    CUPY_AVAILABLE = False

class HyperdimensionalNavigator:
    def __init__(self, dimension: int = 4):
        self.dimension = dimension
        self.quaternion_basis = np.array([
            [1, 0, 0, 0],  # 1
            [0, 1, 0, 0],  # i
            [0, 0, 1, 0],  # j
            [0, 0, 0, 1]   # k
        ])

    def quaternion_trajectory(self, coords: np.ndarray) -> complex:
        """
        Computes quaternion-based trajectory probability.
        Uses hyperdimensional algebra for path optimization.
        """
        if len(coords) < 2:
            return 1.0 + 0.0j  # Return 1 for single point
        
        # Convert to quaternion representation
        quaternions = []
        for coord in coords:
            q = np.zeros(4)
            for i, val in enumerate(coord[:4]):
                q[i] = val
            quaternions.append(q)
        
        # Compute path integral using quaternion multiplication
        result = 1.0 + 0.0j
        for i in range(len(quaternions) - 1):
            q1 = quaternions[i]
            q2 = quaternions[i + 1]
            
            # Quaternion multiplication
            w1, x1, y1, z1 = q1
            w2, x2, y2, z2 = q2
            
            w = w1*w2 - x1*x2 - y1*y2 - z1*z2
            x = w1*x2 + x1*w2 + y1*z2 - z1*y2
            y = w1*y2 - x1*z2 + y1*w2 + z1*x2
            z = w1*z2 + x1*y2 - y1*x2 + z1*w2
            
            # Add to path integral (avoid zero multiplication)
            if abs(w) < 1e-10 and abs(x) < 1e-10 and abs(y) < 1e-10 and abs(z) < 1e-10:
                result *= (1.0 + 0.1j)  # Small non-zero value
            else:
                result *= (w + x*1j + y*1j + z*1j)
        
        # Ensure result is non-zero
        if abs(result) < 1e-10:
            result = 0.1 + 0.1j
        
        return result

    def optimize_path(self, start: np.ndarray, end: np.ndarray, 
                     constraints: dict | None = None) -> np.ndarray:
        """
        Optimizes navigation path using hyperdimensional constraints.
        """
        # Simple linear interpolation for now
        # In a full implementation, this would use advanced optimization
        steps = 10
        path = np.linspace(start, end, steps)
        return path

    def compute_metric_fluctuation(self, position: np.ndarray) -> float:
        """
        Computes local metric fluctuations for navigation.
        """
        # Simulate metric fluctuations based on position
        fluctuation = float(np.sum(position**2) * 1e-6)
        return fluctuation 