Higgs Module/requirements.txt
numpy==1.26.4
scipy==1.14.1
cupy-cuda12x==13.3.0
pytest==8.3.3